refactor: showRestaurantDetailModal 함수 모듈화
